+++ defInt8.go
@@ -43,9 +43,9 @@ func (def *defInt8) value(offset int) (value interface{}, err error) {
 	on := def.ociNumber[offset]
 	r := C.OCINumberToInt(
 		def.rset.stmt.ses.srv.env.ocierr, //OCIError              *err,
-		&on,                        //const OCINumber       *number,
+		&on,                         //const OCINumber       *number,
 		byteWidth8,                 //uword                 rsl_length,
-		C.OCI_NUMBER_SIGNED,        //uword                 rsl_flag,
+		C.OCI_NUMBER_SIGNED,         //uword                 rsl_flag,
 		unsafe.Pointer(&int8Value)) //void                  *rsl );
 	if r == C.OCI_ERROR {
 		err = def.rset.stmt.ses.srv.env.ociError()