+++ doc.go
@@ -716,7 +716,9 @@ nullable Go types by default:
 		SetClob(ora.OraS).
 		SetBlob(ora.OraBin).
 		SetRaw(ora.OraBin).
-		SetLongRaw(ora.OraBin)
+		SetLongRaw(ora.OraBin).
+		SetFetchLen(100).
+		SetLOBFetchLen(100)
 	sc.StmtCfg = cfg
 	srv, err := env.OpenSrv(sc)
 	// any new SesCfg.StmtCfg, StmtCfg.Cfg will receive this StmtCfg
@@ -764,6 +766,10 @@ don't have a unique Go type.
 The default for SELECTing [BC]LOB columns is a safe Bin or S,
 which means all the contents of the LOB is slurped into memory and returned
 as a []byte or string.
+
+The DefaultLOBFetchLen says LOBs are prefetched only a minimal way, to minimize
+extra memory usage - you can override this using
+`stmt.SetCfg(stmt.Cfg().SetLOBFetchLen(100))`.
 
 If you want more control, you can use ora.L in Prep, Qry or
 `ses.SetCfg(ses.Cfg().SetBlob(ora.L))`. But keep in mind that Oracle restricts