+++ util_arr.go
@@ -26,7 +26,7 @@ type ociDef struct {
 }
 
 func (d *ociDef) defineByPos(position int, valuep unsafe.Pointer, valueSize int, dty int) error {
-	d.ensureFetchLength(MaxFetchLen)
+	d.ensureFetchLength(d.rset.stmt.Cfg().FetchLen())
 	// If you omit the rlenp parameter of OCIDefineByPos(), returned values are blank-padded to the buffer length, and NULLs are returned as a string of blank characters. If rlenp is included, returned values are not blank-padded. Instead, their actual lengths are returned in the rlenp parameter.
 	if r := C.OCIDEFINEBYPOS(
 		d.rset.ocistmt,    //OCIStmt     *stmtp,
@@ -63,6 +63,9 @@ var (
 )
 
 func (a *arrHlp) ensureFetchLength(length int) {
+	if length <= 0 || length >= MaxFetchLen {
+		length = MaxFetchLen
+	}
 	a.Lock()
 	defer a.Unlock()
 	if cap(a.nullInds) >= length {